orchestrator/requirements.txt
flask
flask[async]
aiohttp
requests
pyyaml
flask-cors
pip3 install GitPython